pyproject.toml
[tool.poetry]
name = "biu-llm-project"
version = "0.1.0"
description = ""
authors = ["IgorDiamandi <161752477+IgorDiamandi@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
datetime = "^5.5"
pydantic = "^2.9.2"
flask = "^3.0.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
scikit-learn = "^1.5.2"
faiss-cpu = "^1.9.0"
transformers = "^4.46.2"
sentence-transformers = "^3.2.1"
python-dotenv = "^1.0.1"
blinker = "^1.8.2"
charset-normalizer = "^3.4.0"
boto3 = "^1.35.55"
jupyter = "^1.1.1"
lab = "^8.3"
jupyterlab = "^4.3.0"
openai = "^1.54.3"
pyyaml = "^6.0.2"
ipykernel = "^6.29.5"
tqdm = "^4.67.0"
anthropic = "^0.39.0"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
pymupdf = "^1.24.13"
pillow = "^11.0.0"
opencv-python = "^4.10.0.84"
numpy = "<2.0.0"
langchain = "^0.3.7"
langchain-community = "^0.3.5"
llama-index = "^0.11.22"
langchain-openai = "^0.2.6"
torch = "^2.5.1"
gradio = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
